mergenetic/pyproject.toml
[project]
name = "mergenetic"
version = "0.1.1"
description = "Evolutionary model merging for LLMs"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "Tommaso Mencattini", email = "tommaso.mencattini@epfl.ch" }, { name = "Adrian R. Minut", email = "robertadrian.minut@gmail.com" }]
keywords = ["llm", "model merging", "evolutionary optimization", "mergekit"]
requires-python = ">=3.10,<3.12"

dependencies = [
  "accelerate==1.3.0",
  "bitsandbytes==0.42.0",
  "einops==0.7.0",
  "fasttext-numpy2",
  "git-python",
  "gradio==5.29.1",
  "lm-eval[vllm]==0.4.8",
  "mergekit==0.1.2",
  "pandas==2.2.2",
  "plotly",
  "psutil==7.0.0",
  "pymoo==0.6.1.3",
  "pyyaml==6.0.1",
  "scikit-learn",
  "transformers==4.45.2",
  "wheel==0.41.2",
  "wrapt==1.16.0",
  "xxhash==3.4.1",
  "yarl==1.9.4",
  "prompt_toolkit==3.0.51",
]

[project.urls]
Homepage = "https://github.com/tommasomncttn/mergenetic"
Repository = "https://github.com/tommasomncttn/mergenetic.git"
Issues = "https://github.com/tommasomncttn/mergenetic/issues"

[build-system]
requires = [
    "setuptools>=61.0", 
    "wheel",
    "numpy>=2.0.2",  # For PyTorch and potentially other build steps
    "torch>=2.0.0",   # If setup.py imports torch for CUDAExtension or other reasons
    "pybind11>=2.6.0" # If your extensions use Pybind11 and setup.py needs it
]
build-backend = "setuptools.build_meta"
# backend-path = ["."] # Optional: default is often sufficient if setup.py is at the root

[tool.isort]
profile = "black"

[tool.setuptools]
license-files = []

[project.optional-dependencies]
    dev = [
      "pytest",
      "flake8",
      "black",
      "isort",
    ]